OpenSearch/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dotenv-python==0.0.1
Events==0.5
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langsmith==0.4.34
ollama==0.6.0
opensearch-py==3.0.0
orjson==3.11.3
packaging==25.0
pydantic==2.12.2
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
zstandard==0.25.0
